Refactor so that we do not pollute job class with our methods
